Fold ToPath into FileOpen, add root benchmarks to keep mix comparable
